+++ CHANGELOG.md
@@ -28,7 +28,7 @@ If you like this project and find it useful, please consider giving it a star on
 
 <a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="120"></a>
 
-## [2.0.16] - Dev branch
+## [2.0.16] - 2025-05-01
 
 ### Changed
 
@@ -48,6 +48,8 @@ If you like this project and find it useful, please consider giving it a star on
 - [eslint]: Remove `eslint-plugin-promise` (not actively maintained) and add optional @typescript-eslint promise rules.
 - [package]: Remove `overrides` that was necessary for eslint-plugin-promise.
 - [eslint]: Add `eslint` v.2.0.0 config.
+- [prettier]: Add `prettier` v.2.0.0 config.
+- [jest]: Add `jest` v.2.0.0 config.
 
 ## [2.0.15] - 2026-04-08